feat!: updated to dotnet 9 plus many other improvements
- We now use pixi to manage the dev env instead
of a devcontainer.

- The latest version of rattler-build is used to
build the conda packages instead of the previous
custom solution

- Runtime dependencies such as icu for
globalisation are included.

- In order to resolve such dependencies at runtime
we have set the LD_LIBRARY_PATH & DYLD_LIBRARY_PATH
environment variables for linux & osx respectively
upon activation.

+++ src/Services/Releases/ReleasesClient.cs
@@ -32,7 +32,7 @@ public class ReleasesClient
         CancellationToken cancel = default
     )
     {
-        var notes = await GetReleaseNotes(version.Major, cancel);
+        var notes = await GetReleaseNotes((int)version.Major, cancel);
 
         ReleaseNoteSdk? release = null;
         if (version.Minor == 0 && version.Patch == 0)
